Adds comments. Cleans up the code
