+++ WeatherCodingChallenge/WeatherCodingChallenge/Extensions/Ext+String.swift
@@ -8,6 +8,8 @@
 import Foundation
 
 extension String {
+    // Simple extension that takes the API date/time string and converts it to look the way I want
+    // In this instance, its set up for the 3-hour Forecast
     func convertToTime() -> String {
         let formatter = DateFormatter()
         formatter.dateFormat = "yyyy-MM-dd HH:mm:ss"
