+++ WeatherCodingChallenge/WeatherCodingChallenge/Model/CityForecast.swift
@@ -9,6 +9,8 @@ import Foundation
 
 struct CityForecast: Codable {
     
+    // This model complies the JSON data into one easy-to-use and object.
+    // Improves readability
     let cityName: String
     let temp: Double
     let feelsLike: Double
