+++ WeatherCodingChallenge/WeatherCodingChallenge/Model/ThreeHourForecast.swift
@@ -9,6 +9,7 @@ import Foundation
 
 struct ThreeHourForecast: Codable {
     
+    // Consolidates all the API JSON data to one simple location, without the need to drill down each time
     let time: String
     let chances: Double
     let conditionsID: Int
